Update index_phase1.rst - metric selection
add clarity on metric use: Select the data type and the metrics which you would like to displayed on the chart. 
Events are shown on the chart only if the impact metric is non-zero in the historical loss catalogue. It is common for events to have a zero loss for Economic Damage and non-zero loss for people affected - therefore the number of events displayed will differ by impact metric selected.

**The metric type shown on this chart when you navigate to the next step (click 'Next'), will be the metric used in the rest of the analysis.**