Update index.rst
The central purpose of this tool is educational, serving as an introduction to the different considerations that go into modelling financial risk within funds. The tool demonstrates how the processes of risk pooling and funding can be structured efficiently and responsibly even with events being highly uncertain year to year. It has been developed to support users to better understand the risk of climate and disaster losses and use of risk pools in response to requests in the development and humanitarian sectors for better understanding of these aspects.